Use longest job time to prevent unnecessary penalizing
Jobs that end before the longest job time should not be penalized because they have certainly NOT exceeded the optimal schedule endtime. The optimal schedule endtime has to be equal to or greater than the longest running job, thus so long as the job ends before that time, we don't care (i.e. those biases should be 0)